scripts_random/requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PySocks==1.7.1
pytesseract==0.3.10
python-decouple==3.6
python-dotenv==1.0.0
requests==2.28.1
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.13
webdriver-manager==3.8.6
wsproto==1.2.0
